Revert "renamed platform to agent_platform trying to get rid of the error in Azure that system attribute is not in platform, I think they are loading the wrong "platform""
This reverts commit 3694b56345d0616ccfd1bacf897d5accea68d618.
